fixed f4560a1 from https://gitee.com/suwenxiang12/graphic_ui/pulls/805 Description: Fallback amplification fontid modification IssueNo: https://gitee.com/openharmony/graphic_ui/issues/I5YRJV Feature or Bugfix: Bugfix Binary Source: No Signed-off-by: suwenxiang <suwenxiang@huawei.com>

+++ frameworks/components/ui_arc_label.cpp
@@ -89,7 +89,7 @@ UITextLanguageDirect UIArcLabel::GetDirect()
     return arcLabelText_->GetDirect();
 }
 
-void UIArcLabel::SetFontId(uint16_t fontId)
+void UIArcLabel::SetFontId(uint8_t fontId)
 {
     InitArcLabelText();
     arcLabelText_->SetFontId(fontId);
@@ -141,7 +141,7 @@ void UIArcLabel::DrawArcText(BufferInfo& gfxDstBuffer,
                            arcLabelText_->GetFontSize(), arcTextInfo, orientation, *style_, opaScale);
 }
 
-Rect UIArcLabel::GetArcTextRect(const char* text, uint16_t fontId, uint8_t fontSize, const Point& arcCenter,
+Rect UIArcLabel::GetArcTextRect(const char* text, uint8_t fontId, uint8_t fontSize, const Point& arcCenter,
                                 int16_t letterSpace, TextOrientation orientation, const ArcTextInfo& arcTextInfo)
 {
     return TypedText::GetArcTextRect(text, fontId, fontSize, arcCenter, letterSpace, orientation, arcTextInfo);
